[FAB-3316] Disable brittle test- TestCloesConn
Was introduced ~ 4 days ago and it seems that it is
too brittle for CI these days.

It fails because it sends a message via a gRPC stream
and expects the sending to return an error
because the stream was supposed to be closed,
but sometimes it isn't.

Disabling now and it will be revisited soon.

Change-Id: I3f6d6861b4b917bacff82ab95d25a9f1738837b1
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>