[FAB-3335] Gossip pull may send zero-length digests
The bug was introduced in a recent commit that added filtering capability.

Basically when the gossip puller gets a hello from a peer,
it iterates over the items it has, and applies a filter on each of them.
If the filter doesn't permit, it skips to the next iteration,
but the returned slice then contains an "empty-string"
digest in the response to the remote peer.

While this isn't dangerous, this causes un-necessary messages
to be sent so this should be fixed.

Also added an optimization that if there are no digests/items to send,
we don't return any response to the sender.

Change-Id: Ic455413dc9e040ea98ef3bb0ee3531f2c6cfc32c
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>