[FAB-3350] Increase test coverage for cryptogen
As we start to package this tool, need to make
sure the test coverage is adequate.  This CR
ups the current test coverage to >90%. It adds
some logic to existing tests and modifies some
functions to use nested if logic as independent
checks for errors were not needed / efficient.

Will need to add coverage for main.go in a follow-up
change.  Also addresses staticcheck errors from
[FAB-3177]

Change-Id: Ia7def11383e7c680ece59e5630cf9debaa6be2b0
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>