[FAB-3316] Reintroduce TestCloseConn back to CI
This commit fixes TestCloseConn that was disabled due to being brittle.
Basically the test creates a stream and authenticates with the
a gossip instance and then makes the gossip instance close the connection
with the stream, and then attempts to send a message via the stream
and see that it fails.

However, due to HTTP/2 buffering, the send sometimes works.
Therefore, I just made the test send ~ 20 messages instead of 1,
hopefully that it would overload the buffer and cause a flush()
in the HTTP/2 implementation and that would fail, which would
make the sending return a failure too.
Also increased the size of the messages that are sent, in order
to make the flushing be more likely triggered

Change-Id: I18695e97f552b582bfa8ee6a3146e1e83a6d1e1d
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>