[FAB-3161] /examples/cluster
The purpose of this patch is to provide a tool that will
demonstrate how to assemble a complete cluster of a
v1.0 architecture system, complete with security.

It currently supports docker-compose (via "make compose")
but future plans including adding "make kubernetes", etc.

Usage
===========

$ make
Usage: make [target]

Supported Targets:

- compose-up [options]:   builds a docker-compose based cluster
                          [options: TLS=[true|false] (default: true)]
- compose-down:           shuts down a docker-compose based cluster
- clean:                  cleans local artifacts and, where applicable,
                          destroys cluster
- help:                   displays this help message

Example:

        make compose-up TLS=false

Prerequisites:

- Ensure you run "make docker" in both fabric.git and fabric-ca.git
  prior to execution

Notes
===================

The resulting cluster is not production-grade, per se, largely
because of the way the keys are generated all in one place
via cryptogen tool.  However, there are some contexts in
which this might be ok anyway (e.g. loading kubernetes
secrets) and at the very least, its a prescription for the
major steps needed.

You may think of this tool as similar to e2e but with slightly
different goals.  E2E is targetted at easily standing up
a fixed configuration and ensuring chaincode may deploy.  This
is targetted at easily standing up a dynamically generated
crypto configuration in a somewhat realistic manner in a few
different environments.

Change-Id: I16561aa252212ca20645fa5b4a1a89ffd70c1a78
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>