[FAB-3428] configtxgen: Report selected configuration
This patch simply adds an INFO level log message to indicate
which configuration was ultimately selected.  This may be
helpful to the user since the process for selecting the
config may be automated by the software.  This could be
confusing if the user believes one configuration was in
effect when the system loaded an alternate.

Fixes FAB-3428

Change-Id: I085c4b719c85a9d7d4af86843f89ac819da9b6e4
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>