[FAB-3473] Improve UT coverage for peer/clilogging
This CR reworks some of the peer/clilogging commands. With these changes
in place and updates to the unit tests, the unit test coverage for
the package improves from 25% to 66%. Further work is still needed to
improve the coverage for this package and other peer CLI packages but
this is at least a start.

Change-Id: I77df0ca7af7565953ea43e287067d530a7a9dc6d
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>