[FAB-3530] Gossip - add block seq# validation
When a peer receives a data message (contains a block) it doesn't verify
that the sequence of the block matches the sequence inside the marshaled block,
and just calls VerifyBlock.

The implementation already unmarshals the block, so there is no point
in un-marshalling it outside.
I extended the interface VerifyBlock to accept also a sequence number.

Change-Id: I85c6c57f26f5e404fe8e7ece2abffa58eaded625
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>