[FAB-3232] Fix chaincode container logging
This CR fixes the logging setup for chaincode containers. Previously,
the code would only set the level for the 'shim' module and all other
loggers in the scope of the container would use the default. With this
change in place, CORE_CHAINCODE_LOGGING_LEVEL can be used to set the
default logging level for all modules within a chaincode container and
CORE_CHAINCODE_LOGGING_SHIM can be used to override the level for the
'shim' logging module.

Change-Id: I019fd3995e2af25a170d751628293fb215a748bf
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>