[FAB-3445] unrecognized characters in the gossip log
Some of our modules print funny characters such as:
2017-04-27 08:21:33.648 UTC [flogging] setModuleLevel -> DEBU 2eea Module 'gossip/gossip#^>+�>�Q�
�N�5�W"���?�xʷ��?J

This is caused because we have (for testing) module names that consist
of both the module itself + a descriptor for the peer instance, since
all instances run in the same process in unit tests.

In some modules the descriptor is the PKI-ID of the peer, which
in tests its simply the endpoint of the peer, but in production
this is the hash of the cert --> which isn't ASCII.

I adjusted our logging config to use only the module in production,
and use the existing flow only in test code, where it belongs.

Change-Id: Idc3ec6a4b9223d1506873a3af7210bf1271654f2
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>