[FAB-3613] JoinChan gossip broken if no anchorPeers
Since https://gerrit.hyperledger.org/r/#/c/7105/ is merged,
when a channel is created- no anchor peers are supplied inside.

There is a bug in gossip_service.go that if a config with an empty
set of anchor peers for each org is given, the channel's membership
(list of orgs) is considered empty, and then each peer considers
itself alone in the channel.

It "recovers" when the update config that does contain the anchor
peers comes, but until then - it doesn't consider even peers of
its own org as members of the channel because of the bug.

I fixed it and added a test that simulates the flow that caused
the bug.

Change-Id: Ibc71747f2fd56add1361e996c1af89c8c3db479b
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>