[FAB-3315] Certificate sanitization
In order to address the FAB, this change-set
does the following:
1. When an msp identity is created, the corresponding
x509 certificate gets sanitized first.
Sanitization means the following: In case of
certificate signed with ECDSA, the sanitization
checks that the signature is in low-S, if this is not
the case, a clone of the certificate is created
with the new signature.

Change-Id: I0ca6938d26dcc98352ccd1b5498a68cd670afdea
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>