[FAB-3698] def inst. policy needs to include channel
Instantiation policy needs to be defined as part of the
install package. Short of that, the default policy needs to
include all the admin MSPs in the the channel instead of just
the Local MSP of the peer on which the instantiation occurs.

This - among other things -  allows any endorser in the channel
to be part of instantiation.

Change-Id: I16d02e168bcc9a28b0156d734e310268769ea1c7
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>