[FAB-3733] Fix the CouchDB option for e2e_cli
Include a docker-compose file that will work as an addition to existing
docker-compose file and will 1) create the necessary CouchDB containers
and 2) include the necessary environment variables in the peers.

Deleting the previous directions in the existing docker-compose and
updating the driver script so the CouchDB option can be run simply by
calling the driver with and added parameter "couchdb".

Also exposed all chaincode examples instead of just example02, because
marbles02 (JSON based) is a better demonstration on CouchDB.

Document update to follow.

Change-Id: I9c4dbb1f08a6e07319077fb79942483a7782980f
Signed-off-by: Adnan Choudhury <adnan.choudhury@itpeoplecorp.com>