[FAB-3744] Gossip: only pull from peers in the same org
Currently, the pull mechanism works with peers from any
orgs that are eligible of being in the channel.

This has the following problem:

If in a certain channel there are 2 orgs: {A, B} and a peer
from orgB initiates a pull with a peer from orgA and as a result,
the peer from orgA sends the peer from orgB blocks with
sequences [n... n+k] it has received from either the ordering
service or from peers in its own org, it is not safe because
a block i in [n.. n+k] can be a configuration block that
evicts orgB from the channel, and as a result,
orgB would receive blocks it isn't eligible of receiving.

I fixed this by checking that the org is the same org as the peer
and also added a test

Change-Id: I348a22334a0751bb09a5f962ddfd08d516c12f30
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>