[FAB-3755] Gossip:Add remote peer endpoint to ConnInfo
In gossip we have the following struct that can be extracted
from a point-to-point message:

type ConnectionInfo struct {
    ID       common.PKIidType
    Auth     *AuthInfo
    Identity api.PeerIdentityType
}

In the logs we write the ID of the peer but it's a byte slice,
and doesn't tell much about the peer.

On the other hand, when a remote peer connects to a peer,
we extract its endpoint anyway.

I added the endpoint to the struct and added an assignment
at the handshake time.

Change-Id: If29abda09f2b066332099a45328c26027641215b
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>