[FAB-3763] Fixing Intermediate CA certs sanitization
This change-set addresses the following issue:
Currently, certificates of intermediate CAs are sanitized in the same
way root CA certificates are. Namely, by using as parent certificate
the certificate itself. But intermediate CA certificate must have
a parent and therefore the sanitation must happen with the respect
to that certificate.

Change-Id: I2f9c7d7eb9e4c1d3aab722ddac5ffdf8189fdf37
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>