[FAB-3368] Zero length CouchDB attachment delays
CouchDB attachments with zero length causes CouchDB to return
successfully but the connection is not available for future
connections.  This is a bug identified in CouchDB.

Zero length attachments are valid.  When an empty byte array
is saved, CouchDB stores the value as a zero length attachment.

The next subsequent client request over the same connection times
out reading headers from server.

This is the CouchDB JIRA item for tracking purposes.
https://issues.apache.org/jira/browse/COUCHDB-3394

The fix in this change checks to see if the attachment in
SaveDoc is zero length.  If the attachment is zero length,
then request close value is set to true.  This setting will
not allow the http connection to be reused and will prevent
connection timeouts.

Change-Id: I1a1a33c44a3b96dfd2bffaf1cec5fe9ece084817
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>