[FAB-3846]: Behave test feature file placeholders
This commits a number of placeholder tests that
will be executed using the behave framework.

This is only a small representation of the tests that
will be written and executed.

Execute these tests by typing:
  $ behave --tags=-skip

or these tests will be executed when executing the
runDailyTestSuite.sh file with the correct tags and
will generate the results in junit with output files
in the test/regression/daily directory according to
the feature file name.

I have also updated the .gitignore to ignore the
xml files that are generated as a result of executing
the tests for CI ingestion.

Change-Id: I61fc34fd430aace182b8bdffe36d42ad73e2cc6f
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>