[FAB-3562] CLI displays unexpected invoke result
This CR improves the log messages after a successful invoke. The
previous log message displayed the ESCC invoke result instead of
the chaincode invoke result, which was confusing to end users. The
previous log message now specifies that it is for the ESCC result
and adds a new log message to display the chaincode invoke result.

This CR also adds a few logging messages to the endorser code that were
helpful during my time debugging and working to understand the
endorsement logic.

Change-Id: I02cb0dfe620e7c17472b8b12467c746c27598942
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>