[FAB-3953] Re-introduce disabled goroutine leak test
The check was disabled in FAB-3257 and deals with checking
if goroutines that are created during a test are dead after the test ends.
It failed because sometimes the number of goroutines that the test started
with, was higher than the number it ended with.

It should be re-introduced.

I think we should simply instead of checking that the number of goroutines
hasn't changed from the start of the test to the end of the test - check
that it didn't increase, because that proves that a goroutine was created
at the time of the test and hasn't finished.

Change-Id: Id8caabff4723699ede6047d2640f4442948e52e4
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>