[FAB-4077] Fix gossip false negative WARN log
When a peer determines whether a given peer is eligible for
receiving blocks for some channel, it looks up the latest state
info message that the remote peer has sent.

In some situations, especially at boot time - the peer might get
a message for that channel (i.e a hello message for blocks pulling)
from that remote peer before it has received a stateInfo message
from that peer. This results in a printing of a warning that a peer
attempted to pull blocks but wasn't eligible for the channel.

This warning is a false negative and disappears after a while when
the peer receives the StateInfo message of the remote peer.

I added a check to ensure that if the peer doesn't have a stateInfo
message of the remote peer in the cache, to abort the check.

Change-Id: I539a600e4eb592eb3e14db8433c6490520618fed
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>