[FAB-4078] StateInfo expiration membership awareness
The stateInfo message store is used in gossip to know the membership
of the peers of the channel.
Each entry in the message store is the latest StateInfo message a peer
has sent either directly or indirectly.
Peers send StateInfo messages in the following manner:
- Upon reception of a block, a new StateInfo message is created and is
propagated to adjacent peers.
- Once in a while, a peer send StateInfo pull requests to adjacent peers
  and they send back the entire StateInfo message set stored in their
  memory that they have received from other peers (reconciliation).

However - the StateInfo message store is an expirable message store -
its messages are expired after a limited time.

This causes the following problem:
If no new blocks are generated in the channel, no new StateInfo messages
are generated as a result, and eventually -
he StateInfo messages of peers are cleaned up from memory.

This causes many problems, namely that when a new peers joins a channel
after no new blocks were created in that channel for a while -
it will not get StateInfo messages from other peers and thus won't be
able to pull blocks from them, unless it is configured to use leader
election, and then it will elect itself as a leader and connect directly
 to the ordering service.

Change-Id: Ifa38a714a448ae3708e9852384b5e4107380e947
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>