[FAB-4105] Fab proto translation methods
With the work done for FAB-4100 through FAB-4104, it is now possible to
add a few small descriptive methods to proto methods, and the proto
translator framework can then transalte the binary proto (including
fields which would ordinarilly be expressed as binary, or nested binary)
into pure binary-less JSON.

This CR adds the methods necessary to annotate the fabric protos related
to configuration.  This now allows the proto translation framework to
print fully human readable versions of both config update messages, as
well as of configuration blocks (and any subset thereof).

This CR appears to introduce substantial amounts of code with very
little tests.  However, the two tests in blackbox_test use every method
introduced.  Additionally, because these methods are only used for proto
printing, they can have no affect on the stability of the fabric
overall.

This is a key component for FAB-1678.

Change-Id: If166a55097dccef5208c1dc2ccd3bf039b7637d0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>