[FAB-4087] Fix test failure in dockercontroller
Test_Start is failing intermittently because the function variable 'err'
in the Start function was used in a go routine function, which was
clobbering the function variable, only in cases the go routine was able
to finish before the err is returned by the Start function.
The fix is to use a local variable in the go routine function.

Change-Id: Ibbf0f4e1b551020554ff0b06c2c16bc0f39dbc60
Signed-off-by: Anil Ambati <aambati@us.ibm.com>