[FAB-4107] Expose proto translator via REST
As part of FAB-4100, there is a requirement to expose configuration
utilities through a REST interface.

This CR adds a proto-translator REST component.  Further documentation
is pending, but the high level usage is:

POST binary-proto to /protolator/decode/fq.MessageType replies with a
deeply marshaled JSON version of the proto.

POST json-doc to /protolator/encode/fq.MessageType replies with the
proto version of the deeply marshaled JSON doc.

Change-Id: I595b92b1c292d8d4d360b0bd4223b138615143ac
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>