[FAB-4167] Expose config update compute via REST
As part of FAB-1678, the tool needs to support computing a config update
based on the contents of two configs.  This CR adds a new REST endpoint
for handling this computation.

It add the target /configtxlator/compute/update-from-configs which takes
a POST of multipart/formdata with two files, the first for the field
'original' and the second for the form 'updated'.  These should both be
proto marshaled common.Config structures. A final 'channel' parameter
may be supplied to set the channel id for the update.

The server then invokes the configtxlator.update package to compute the
config update based on the inputs, and returns it as a marshaled
common.ConfigUpdate message.

Change-Id: I0ca13d6de41da8b0ce9bfc66685476b42a6991e6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>