[FAB-3358] Remove testchainid from peer startup
There is currently a --peer-defaultchain option
which can be passed to peer node start which will
setup a testchain.  This was used in the past while
work was underway to support creating channels but
at this point it is no longer needed and we should not
have test code in the production path.

This is Part 1 of 2 to fix this.  First, we remove
the actual code for creating testchainid but we
leave the CLI flag because the SDK e2e's have it
in their compose cmd.

Change-Id: I7229d647976a5d927dc0c53cfaeebd7f5fa9c811
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>