[FAB-4265] Fix overloaded peer channel create cmd
The peer channel create command is currently being used in the e2e tests
to perform a configuration update.  This is very unintuitive and gives
the wrong impression to anyone looking at the e2e example.

This CR simply adds an 'update' command to peer channel to make the
operation more clear.

Change-Id: I33244affb5cd013ca878123b2bd7c8b2c00690b6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>