[FAB-4305] Implement gRPC keepalive support
There are a couple of defects related to
long running gRPC connections being
broken by intermediaries due to inactivity.
The latest version of the gRPC libraries
added support for keepliaves to deal with
this situation.

This change adds keepalive support for
gRPC server and client connections.  The
defaults are currently being used and no
user configuration has been exposed.
The defaults are sane (5min inactivity
and 20sec ping response) for typical
environments.  Will submit a separate
change for gRPC configuration in general,
although we can decide whether or not to
take that on at this point.

Change-Id: Idf63e37eb219e7961a2217193d6c00a69d259998
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>