[FAB-4340] Remove TravisCI_Readme.md
I don't understand the reason for this file's existence. Isn't a link to
https://docs.travis-ci.com/user/getting-started/ somewhere in our Wiki
enough? This changeset removes the file.

Note that the text on it — if it were to survive for some reason — needs
serious revising as well. Formatting, capitalization, and punctuation
are all over the place. There are also references to "openchain" and the
project is referred to as "hyperledger".

Change-Id: Icf1147e28eb5d0e4330d547b7766f9ea31cdde76
Signed-off-by: Kostas Christidis <kostas@christidis.io>