[FAB-4347] prop resp to return error on cc error
CC responses that are greater than an threshold (currently 500, soon
to be 400) will not be endorsed. As the error is from chaincode and not
directly from fabric, endorser returns error. However it makes sense to
return a special 500 error with message, "chaincode error" to signify that
the response was not endorsed.

SDK and users could deep inspect inner response from chaincode for more
information.

Change-Id: If0ba17908576791bd1933bec17a65db8698f8aef
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>