[FAB-4373] Fix orderer system channel Admins
The default /Channel/Admins policy requires the MAJORITY of sub Admins
policies to be satisfied, for the /Channel/Admins to be satisfied.
Usually this means both of /Channel/Orderer/Admins and
/Channel/Application/Admins.  However, on the ordering system channel,
this means both of /Channel/Orderer/Admins and
/Channel/Consortiums/Admins.  Because the Consortiums group does not
have an Admins policy defined, the /Channel/Admins policy can never
evaluate to true.

This CR defines an Admins policy at the /Channel/Consortiums/Admins
level which accepts all signature sets.  This causes the /Channel/Admins
policy to become equivalent to the /Channel/Orderer/Admins in the
ordering system channel case.  The /Channel/Consortiums/Admins policy is
not used for any other purpose, so the accept all definition has no
other implications.

Change-Id: Ib8a247743f52d9d4bc5b5c80d8351a91647d3f6c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>