[FAB-4420] Suppress benign channel create warning
During channel creation, the orderer creates a template channel to apply
the channel creation request to.  This allows for the same signature
checking and validation path to be re-used that is used for config
updates.

However, the template config which the update is applied to is not a
truly valid 'fabric application config', because it does not define
/Channel/Application/Readers or /Channel/Application/Writers, although
it does define an application section.

This causes warnings to superfluously be printed suggesting that the
current config is likely incorrect.

This CR makes an admittedly hacky, but minimally invasive change to
suppress these log messages.

Change-Id: I06dc42a4185b1bf585869a8003436dc622e4606b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>