[FAB-4431] Orderer multichain api consenter errors
The orderer multichain consenter API currently does not provide a
mechanism to provide feedback for a channel when the consenter is out of
sync or is otherwise not producing data for that channel.

This CR adds an Errored() method to the common consenter Chain
interface, and consumes that API via the Deliver common code.

It provides only a skeleton implementation for the consenter
implementatinos (Kafka and Solo).  The Kafka implementation will provide
a more robust implementation of the API in the near future.

Change-Id: I22863f2a8b37932a5ecbd5241372e2a5c512571d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>