[FAB-4427] Fix unchecked errors in configtxgen
There was an unchecked file read in configtxgen which caused missing
files to be ignored when read.

This ordinarilly would not have resulted in a crash, but additionally,
the envelope extract utility method did not have a nil check on the
block data, which caused a crash.

This CR fixes both of these issues (adding tests for each).

Change-Id: Ie21b397f86a9c88da4b488c81f3f82df8b2ab3c7
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>