[FAB-4463] Fix nil dereference in cauthdsl
When a specially malformed config transaction is submitted, it is
possible to trick cauthdsl into attempting to compile a nil signature
policy.  This could result in a crash.

This CR checks for a nil signature policy, and adds a test case for this
condition.  It also slightly enhances the logging in the package.

Change-Id: I78194e13fc94c2715db169121dbff9ea4387f69b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>