[FAB-4468] Create configtx sanity check code
The configtx processing framework does checks on config to make sure
that it is well formed, authorized, etc.  However, there are certain
configurations mistakes which are not considered fatal, but which still
usually, do not make sense.

This CR creates a library for doing configtx sanity checking.  At the
moment, it checks to make sure that the normal configtx processing
framework considers it well formed, then checks to make sure all
policies of type Signature policy only refer to MSP principals dependant
on orgs defined within the system (where possible).

The Checking function returns a set of messages, including General
errors, and Element Errors, and Element Warnings.  The Element types
include a jquery type path from the channel group to the offending
configuration element, along with a message describing what the error or
warning is.

Additional sanity checks may be defined in the future.

Change-Id: I944c6dc0378523ef17f11c2b4b46adfbca38d6d8
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>