[FAB-4469] Add REST endpoint for sanity check
The sanity checking code added in FAB-4468 currently has no external
interface.  In order to satisfy the requirements of FAB-3831, it is
necessary to expose an external interface to this sanity checking
ability.

This CR adds a REST endpoint at /configtxlator/sanitycheck/config which
the user may POST a marshaled common.Config to.  The REST target then
replies with a JSON document, indicating the error types and paths, as
reported by the sanity checking code.

Change-Id: I97616eec67aebd77ea8191b8ae39808716330aec
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>