[FAB-4083] Fix filesize-related defaults for orderer
The `Producer.MaxMessageBytes` default we were setting on the Kafka
producers was incorrectly set to Kafka's default value for
`socket.request.max.bytes`. (I actually set it to a value that was just
a byte shorter than this when I pushed in the refactor changeset a few
days ago in order to silence the sarama logs, and added a FIXME note for
this very reason.) This value is incorrect because it doesn't account
for headers. This changeset assumes a too-generous 1MB for headers and
decreates `Producer.MaxMessageBytes` accordingly.

This changeset also updates the default value for AbsoluteMaxBytes in
`configtx.yaml`. The current value is incorrect because it doesn't
account for the wrapping of messages in the container `KafkaRegular`
format. At a minimum, it should be decreased compared to
`Producer.MaxMessageBytes` by an equal padding. This changeset does this
for all `configtx.yaml` files currently in our codebase.

Finally, it sets the default, suggested value for AbsoluteMaxBytes in
the sampleconfig `configtx.yaml` file to a much more reasonable 10 MiB,
well within the constraints we outlined above. The default check in
`localconfig` has been updated accordingly.

See the JIRA item for a detailed explanation of the issue.

Change-Id: I8821982707282c5069c1a66dd2a121b894b3a3c8
Signed-off-by: Kostas Christidis <kostas@christidis.io>