[FAB-4538] Disconnect deliver clients after revoke
If a client of the Deliver interface requests to seek until max_uint64,
then this client will stay connected receiving blocks indefinitely.

If that client's access is revoked, either by removing the organization
from the channel, or the client's certificate is revoked, then this
client is not disconnected, and will continue to receive blocks
indefinitely.

This CR causes the Deliver code to check the configuration sequence
number before each block is delivered.  If the configuration sequence
has been incremented, then the client's authorization is checked once
more and ejected if the new configuration does not permit the client's
request.

Change-Id: Ie852a9a8d435917ef1e7dce2025122c791fc9248
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>