[FAB-4560] Config updates with no effect validate
The config update logic validates that a config update is not empty, and
that any modifications made are validly signed.

However, it is possible that if the config update is not empty, but it
has no effect on the configuration (ie, the read and write sets are
equal), then the configuration will still apply, without any signatures,
because there are no modifications made to be validated.

This CR adds an explicit check which disallows update sets which do not
modify at least on configuration element.

Change-Id: If2be92aaf3a917ac7dbd617226594331d619e09d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>