[FAB-3555] Peer panic on SIGSEV connecting to CouchDB
A peer panic was encountered connecting to CouchDB on
peer startup.   This occurred in the code after multiple
retries.

The panic was caused by a null pointer when attempting
to read the status code from an http return object.  The
http return object was already checked for an error.

The situation causing the null pointer can only be caused
if the golang http return has a nil for both the response
and the error.   According to golang spec this is not
possible.

An additional check is being made to ensure a good error
message is returned from the CouchDB connection request.

Change-Id: I13740a32231d49faf3a6acf7b37f7bc6e4bba3fa
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>