[FAB-4621] Modify channel names in tests
With FAB-2487 in, uppercase characters are no longer allowed in channel
names. This doesn't affect the channel names we use in our tests since
they do not go through the validate function that was modified in
FAB-2487, but we should give them valid names nonetheless for
consistency.

This changeset does the necessary modifications.

Change-Id: Id3d02138b5c767ad1fd560e387caaadd6bafa2d6
Signed-off-by: Kostas Christidis <kostas@christidis.io>